+++ manual/source/version_history.rst
@@ -28,6 +28,11 @@ to maintain the expected high quality of the code.
   into the simulation and all particles are sent back. The :code:`protonsAndIonsOnly`
   Boolean flag still works, but by default is false. Any linked tracking code must now
   filter the particles they can handle themselves.
+* The run level histograms are no longer 'simple histograms' but are now per-event
+  average histograms that are calculated during the simulation. By default, 3D scoring
+  meshes are no longer stored per-event and only the run-level per-event averages are
+  stored.
+
 
 New Features
 ------------
@@ -39,6 +44,8 @@ New Features
 
 * New :code:`CreateEmptyBdskimFile` in the :code:`DataDummyClass` to give this function in pybdsim
   for easy customised skimming in Python.
+* Run histograms are now per-event average histograms as opposed to simple histograms.
+* rebdsim is now faster as looping over data is only done if required.
 
 **Bunch Distributions**
 
@@ -94,6 +101,15 @@ New Options
 |                                     | the design rigidity for normalised fields             |
 |                                     | accordingly.                                          |
 +-------------------------------------+-------------------------------------------------------+
+| storeEventLevelHistograms           | Default on. If on, 1D and 2D histograms are stored    |
+|                                     | with each event with that event data only. Applies to |
+|                                     | all possible 1D and 2D histograms.                    |
++-------------------------------------+-------------------------------------------------------+
+| storeEventLevelMeshes               | Default off. If on, 3D and 4D meshes are stored with  |
+|                                     | each event with that event data only. Incurs a very   |
+|                                     | large file size.                                      |
++-------------------------------------+-------------------------------------------------------+
+
 
 General Updates
 ---------------
@@ -166,6 +182,7 @@ Output Changes
   an element (:code:`staEk`) have all been added to the model tree in the output as
   calculated by BDSIM as it now integrates the time and acceleration / decceleration
   along the beamline.
+* Run histograms are now per-event averages.
 
 
 Output Class Versions
@@ -178,7 +195,7 @@ Output Class Versions
 +===================================+=============+=================+=================+
 | BDSOutputROOTEventAperture        | N           | 1               | 1               |
 +-----------------------------------+-------------+-----------------+-----------------+
-| BDSOutputROOTEventBeam            | N           | 6               | 7               |
+| BDSOutputROOTEventBeam            | Y           | 6               | 7               |
 +-----------------------------------+-------------+-----------------+-----------------+
 | BDSOutputROOTEventCavityInfo      | N           | 1               | 1               |
 +-----------------------------------+-------------+-----------------+-----------------+
@@ -200,7 +217,7 @@ Output Class Versions
 +-----------------------------------+-------------+-----------------+-----------------+
 | BDSOutputROOTEventModel           | Y           | 6               | 7               |
 +-----------------------------------+-------------+-----------------+-----------------+
-| BDSOutputROOTEventOptions         | N           | 8               | 8               |
+| BDSOutputROOTEventOptions         | Y           | 8               | 9               |
 +-----------------------------------+-------------+-----------------+-----------------+
 | BDSOutputROOTEventRunInfo         | N           | 3               | 3               |
 +-----------------------------------+-------------+-----------------+-----------------+